refactor(bot): split debug command handlers
